Refactor the reset password form
