+++ todoqueue_frontend/src/App.js
@@ -13,10 +13,10 @@ import ForgotPassword from './components/forgotPassword';
 import Login from './components/Login';
 import { Logout } from './components/logout';
 import { fetchHouseholds } from './api/households';
-
 import { Navigation } from './components/navigation';
+import { ResetPassword } from './components/resetPassword';
+
 import { SignUp } from './signup';
-import { ResetPassword } from './resetPassword';
 import { ManageHouseholds } from './households';
 import UserStatistics from './UserStatistics';
